refactor: Improve member system message rendering
